gcp/requirements.txt
# Function dependencies, for example:
# package>=version
pymongo>=4.3.3
requests>=2.28.1
pandas
numpy
python-telegram-bot
scikit-learn
emoji